+++ libs/graphics/include/graphics/systems/animation.h
@@ -8,9 +8,9 @@
 namespace graphics::systems::animation
 {
 
-	void update_flash(entt::registry& reg, double dt);
-	void update_shake(entt::registry& reg, double dt);
-	void update_shake_once(entt::registry& reg, double dt);
+	void update_flash(entt::registry& reg, float dt);
+	void update_shake(entt::registry& reg, float dt);
+	void update_shake_once(entt::registry& reg, float dt);
 	void update_shake_base_world(entt::registry& reg, entt::entity e, const glm::mat4& world_matrix);
 	
 }
